32-bit compilation warning fixes windows 7 32-bit IOCTL_SCSI_GET_ADDRESS failover

+++ block_hasher.hh
@@ -41,7 +41,7 @@ public:
 			}
 		}
 
-		for(; data_end - data >= _blockSize; data += _blockSize)
+		for(; (uint32_t)(data_end - data) >= _blockSize; data += _blockSize)
 			Update(data);
 
 		_tail.insert(_tail.end(), data, data_end);
